v1.0

+++ src/products/products.service.ts
@@ -1,11 +1,15 @@
-import { forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, ObjectId, UpdateQuery, } from 'mongoose';
+import { FilterQuery, Model, UpdateQuery, } from 'mongoose';
 import { CategoriesService } from 'src/categories/categories.service';
+import { Category, CategoryDocument } from 'src/categories/schemas/category.schema';
 import { FilesService } from 'src/files/files.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
-import { Product, ProductDocument } from './schemas/product.schema';
+import { Product, ProductDocument, Size } from './schemas/product.schema';
+import * as mongoose from 'mongoose'
+import { capitalize } from 'src/utils/textFormatters';
+import { OrderItem, OrderItemDocument } from 'src/orders/schemas/order-item.schema';
 
 export interface IFilters{
     size?: string
@@ -13,13 +17,6 @@ export interface IFilters{
     price?: string
     sort?: string
     sale?: string
-}
-
-interface IFiltersQuery{
-    sizes?: object
-    color?: object
-    price?: object
-    comparePrice?: object
 }
 
 export interface IPath{
@@ -38,39 +35,68 @@ export interface IProductsResponse{
 @Injectable()
 export class ProductsService {
     constructor(
-        @InjectModel(Product.name) private productModel: Model<ProductDocument>, 
+        @InjectModel(Product.name) private productModel: Model<ProductDocument>,
+        @InjectModel(OrderItem.name) private orderItemModel: Model<OrderItemDocument>,
         @Inject(forwardRef(() => CategoriesService)) private categoriesService: CategoriesService,
         private filesService: FilesService 
         ){}
 
     async createProduct(dto: CreateProductDto, image: Express.Multer.File, media: Express.Multer.File[]){
+        const category = await this.categoriesService.get(dto.category)
+        if(!category) {throw new NotFoundException('Category not found')}
+
         const imageId = await this.filesService.uploadFile(image)
         let mediaIds = []
         if(media){
             mediaIds = await Promise.all(media.map(file => this.filesService.uploadFile(file)))
         }
-        const product = await this.productModel.create({...dto, image: imageId, media: mediaIds})
+        const pathString = category.path.map(x => capitalize(x)).join('/')
+        const product = await this.productModel.create({...dto, image: imageId, media: mediaIds, path: category.path, pathString})
+        await this.categoriesService.linkProduct(category._id, product._id)
         return product
     }
 
-    async get(id: ObjectId): Promise<ProductDocument>{
+    async get(id: mongoose.Types.ObjectId): Promise<ProductDocument>{
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            throw new NotFoundException(`Product with id ${id} doesn't exist`)
+        }
         return await this.productModel.findOne({_id: id, enable: true}).populate(['category', 'image', 'media'])
     }
 
-    async getAny(id: ObjectId): Promise<ProductDocument>{
+    async getAny(id: mongoose.Types.ObjectId): Promise<ProductDocument>{
         return await this.productModel.findById(id).populate(['category', 'image', 'media'])
     }
 
-    async getAll(): Promise<ProductDocument[]>{
-        return await this.productModel.find().populate(['image', 'category'])
+    async getRelatedProducts(id: mongoose.Types.ObjectId): Promise<ProductDocument[]>{
+        const product = await this.productModel.findById(id)
+        const relatedProducts = await this.productModel.find({_id: {'$in': product.relatedProducts}})
+        .populate('image')
+        return relatedProducts
+    }
+
+    async getAll(filter?: string): Promise<ProductDocument[]>{
+        let options = {}
+        if(filter){
+            options = {'$or': [
+                {title: {'$regex': /filter/, '$options': 'i'}},
+                {sku: {'$regex': /filter/, '$options': 'i'}},
+                {id: {'$regex': /filter/, '$options': 'i'}}
+            ]}
+        }
+        return await this.productModel.find(options).populate(['image'])
     }
 
     async getByCategory(path: IPath, filters: IFilters): Promise<IProductsResponse>{
-        let queryOptions = {enable: true}
-        const categoryId = await this.categoriesService.getCategoryIdByPath(path)
-        const {products: productIds} = await this.categoriesService.parseChildren(categoryId)
+        let queryOptions: FilterQuery<ProductDocument> = {enable: true}
+        queryOptions['path.0'] = path.categoryName
+        if(path.groupName){
+            queryOptions['path.1'] = path.groupName
+        }
+        if(path.subGroupName){
+            queryOptions['path.2'] = path.subGroupName
+        }
         const filtersQuery = this.formFilterQuery(filters)
-        queryOptions['_id'] = {'$in': productIds, ...filtersQuery}
+        queryOptions = {...queryOptions, ...filtersQuery}
 
         let query = this.productModel.find(queryOptions)
          .populate([
@@ -103,19 +129,28 @@ export class ProductsService {
         return {count: products.length, products, minPrice, maxPrice}
     }
 
-    async update(id: ObjectId, dto: UpdateProductDto, image: Express.Multer.File, media: Express.Multer.File[]): Promise<ProductDocument>{
+    async update(id: mongoose.Types.ObjectId, dto: UpdateProductDto, image: Express.Multer.File | undefined, media: Express.Multer.File[]): Promise<ProductDocument>{
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            throw new NotFoundException(`Product with id ${id} doesn't exist`)
+        }
         const product = await this.productModel.findById(id)
         if(!product) throw new NotFoundException('Product not found')
-        if(dto.category && product.category !== dto.category){
-            await this.categoriesService.linkProduct(dto.category, id)
-        }
         const updateOptions: UpdateQuery<ProductDocument> = {'$set': {...dto}}
+        if(dto.category && product.category !== dto.category){
+            await this.categoriesService.unlinkProduct(product.category, id)
+            await this.categoriesService.linkProduct(dto.category, id)
+            const category = await this.categoriesService.get(dto.category)
+            updateOptions['$set'].path = category.path
+            updateOptions['$set'].pathString = category.path.map(x => capitalize(x)).join('/')
+        }
         if(image){
+            console.log('image', image)
             await this.filesService.removeFile(product.image)
             const newImage = await this.filesService.uploadFile(image)
             updateOptions['$set'].image = newImage
         }
         if(media){
+            console.log('media', media)
             const newMedia = await Promise.all(media.map(file => this.filesService.uploadFile(file)))
             updateOptions['$push'] = {media: {'$each': newMedia}}
         }
@@ -127,44 +162,58 @@ export class ProductsService {
         return await this.productModel.findByIdAndUpdate(id, updateOptions, {new: true})
     }
 
-    async delete(id: ObjectId){
+    async delete(id: mongoose.Types.ObjectId){
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            throw new NotFoundException(`Product with id ${id} doesn't exist`)
+        }
         const product = await this.productModel.findById(id)
+        const orderItems = await this.orderItemModel.find({product: id})
+        if(orderItems){
+            throw new BadRequestException('You can\'t delete product before all order items with this product are removed')
+        }
         await this.filesService.removeFile(product.image)
         await Promise.all(product.media.map(file => this.filesService.removeFile(file)))
         await product.remove()
+        await this.categoriesService.unlinkProduct(product.category, id)
         return product._id
     }
 
-    async deleteMany(ids: ObjectId[]){
+    async deleteMany(ids: mongoose.Types.ObjectId[]){
         return await Promise.all(ids.map(id => this.delete(id)))
     }
 
-    async assignToCategory(products: ObjectId[], category: ObjectId){
-        return await this.productModel.updateMany({_id: {'$in': products}}, {'$set': {category}})
+    async assignToCategory(products: mongoose.Types.ObjectId[], category: CategoryDocument){
+        return await this.productModel.updateMany({_id: {'$in': products}}, {'$set': {category: category.id, path: category.path}})
     }
 
-    async clearCategory(products: ObjectId[]){
-        return await this.productModel.updateMany({_id: {'$in': products}}, {'$unset': {category: 1}})
+    async clearCategory(products: mongoose.Types.ObjectId[]){
+        return await this.productModel.updateMany({_id: {'$in': products}}, {'$unset': {category: 1}, 'path': [], 'pathString': 'No category'})
     }
 
-    async enableProducts(products: ObjectId[]){
+    async enableProducts(products: mongoose.Types.ObjectId[]){
         return await this.productModel.updateMany({_id: {'$in': products}}, {enable: true})
     }
 
-     async disableProducts(products: ObjectId[]){
+     async disableProducts(products: mongoose.Types.ObjectId[]){
         return await this.productModel.updateMany({_id: {'$in': products}}, {enable: false})
     }
 
-    async getProductsByIds(ids: ObjectId[]){
+    async getProductsByIds(ids: mongoose.Types.ObjectId[]){
         return await this.productModel.find({_id: {'$in': ids}})
     }
 
-    async addRelatedProducts(id: ObjectId, products: ObjectId[]){
+    async addRelatedProducts(id: mongoose.Types.ObjectId, products: mongoose.Types.ObjectId[]){
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            throw new NotFoundException(`Product with id ${id} doesn't exist`)
+        }
         await this.productModel.findByIdAndUpdate(id, {'$push': {relatedProducts: {'$each': products}}})
         await this.productModel.updateMany({_id: {'$in': products}}, {'$push': {relatedProducts: id}})
     }
 
-    async removeRelatedProducts(id: ObjectId, products: ObjectId[]){
+    async removeRelatedProducts(id: mongoose.Types.ObjectId, products: mongoose.Types.ObjectId[]){
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            throw new NotFoundException(`Product with id ${id} doesn't exist`)
+        }
         await this.productModel.findByIdAndUpdate(id, {'$pull': {relatedProducts: {'$in': products}}})
         await this.productModel.updateMany({_id: {'$in': products}}, {'$pull': {relatedProducts: id}})
     }
@@ -178,9 +227,9 @@ export class ProductsService {
         return {minPrice, maxPrice}
     }
 
-    private formFilterQuery(filters: IFilters): IFiltersQuery{
+    private formFilterQuery(filters: IFilters): FilterQuery<ProductDocument>{
         const {size, color, price, sale} = filters
-        let filtersQuery: IFiltersQuery = {}
+        let filtersQuery: FilterQuery<ProductDocument> = {}
         //all filters have format of '[a b c ...]', except '[a b]' for price
         if(price  && /^\[\d+(\.\d+)?(\s\d+(\.\d+)?)?\]$/.test(price)){
             //cutting of '[' and ']' braces
@@ -194,9 +243,9 @@ export class ProductsService {
         }
 
         if(size && /\[(xs|s|m|l|xl|xxl)(\s(xs|s|m|l|xl|xxl))*\]/i.test(size)){
-            let sizes = size.slice(1, -1).split(' ')
+            let sizes = size.slice(1, -1).split(' ') 
             sizes = sizes.map(x => x.toLowerCase())
-            filtersQuery['sizes'] = {'$in': sizes}
+            filtersQuery['sizes'] = {'$in': sizes as Size[]}
         }
 
         if(color && /\[[A-Za-z]+(\s[A-Za-z]+)*\]/.test(color)){
